+++ certbot-dns-cloudxns/certbot_dns_cloudxns/dns_cloudxns_test.py
@@ -6,8 +6,7 @@ import unittest
 import mock
 from requests.exceptions import HTTPError, RequestException
 
-from certbot.plugins import dns_test_common
-from certbot.plugins import dns_test_common_lexicon
+from certbot.tests.plugins import dns_test_common, dns_test_common_lexicon
 from certbot.tests import util as test_util
 
 DOMAIN_NOT_FOUND = Exception('No domain found')
