+++ certbot-dns-linode/certbot_dns_linode/dns_linode_test.py
@@ -5,8 +5,7 @@ import unittest
 
 import mock
 
-from certbot.plugins import dns_test_common
-from certbot.plugins import dns_test_common_lexicon
+from certbot.tests.plugins import dns_test_common, dns_test_common_lexicon
 from certbot.tests import util as test_util
 
 TOKEN = 'a-token'
