+++ certbot-dns-rfc2136/certbot_dns_rfc2136/dns_rfc2136_test.py
@@ -9,8 +9,8 @@ import dns.tsig
 import mock
 
 from certbot import errors
-from certbot.plugins import dns_test_common
-from certbot.plugins.dns_test_common import DOMAIN
+from certbot.tests.plugins import dns_test_common
+from certbot.tests.plugins.dns_test_common import DOMAIN
 from certbot.tests import util as test_util
 
 SERVER = '192.0.2.1'
