+++ certbot-nginx/certbot_nginx/tests/tls_sni_01_test.py
@@ -9,7 +9,7 @@ from acme import challenges
 from certbot import achallenges
 from certbot import errors
 
-from certbot.plugins import common_test
+from certbot.tests.plugins import common_test
 from certbot.tests import acme_util
 
 from certbot_nginx import obj
