+++ certbot/tests/plugins/disco_test.py
