+++ certbot/tests/plugins/manual_test.py
