+++ certbot/tests/plugins/storage_test.py
